nuevo arreglo

+++ src/main/java/com/example/productos/repository/ProductoRepository.java
@@ -3,10 +3,11 @@ package com.example.productos.repository;
 import com.example.productos.models.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 public interface ProductoRepository extends JpaRepository<Producto, Integer> {
     Producto findById(int id);
     ArrayList<Producto> findByNombre(String nombre);
-    ArrayList<Producto> findByPrecio(float precio);
+    ArrayList<Producto> findByPrecio(BigDecimal precio);
 }
